Clean up unused components and share logic
