+++ package.json
@@ -23,11 +23,9 @@
     "react": "19.0.0",
     "react-native": "0.79.2",
     "react-native-gesture-handler": "~2.24.0",
-    "react-native-keyboard-aware-scroll-view": "^0.9.5",
     "react-native-reanimated": "~3.17.4",
     "react-native-safe-area-context": "5.4.0",
     "react-native-screens": "~4.10.0",
-    "react-native-swipe-list-view": "^3.2.9",
     "tailwindcss": "^3.4.17",
     "expo-notifications": "~0.31.2",
     "@react-native-community/datetimepicker": "8.3.0"
