+++ utils/menuDefaults.ts
@@ -1,0 +1,8 @@
+import { ListItem } from "../context/ListsContext";
+
+export const baseMenu: ListItem[] = [
+  { key: "mijnDag", icon: "person", label: "Mijn dag", count: null },
+  { key: "belangrijk", icon: "star-outline", label: "Belangrijk", count: null },
+  { key: "gepland", icon: "event-available", label: "Gepland", count: null },
+  { key: "taken", icon: "check-circle-outline", label: "Taken", count: null },
+];
